1.5.6c

+++ UPLOAD/includes/modules/column_right.php
@@ -6,7 +6,7 @@
  * @copyright Copyright 2003-2019 Zen Cart Development Team
  * @copyright Portions Copyright 2003 osCommerce
  * @license https://www.zen-cart-pro.at/license/3_0.txt GNU General Public License V3.0
- * @version $Id: column_right.php 729 2011-08-09 15:49:16Z hugo13 $
+ * @version $Id: column_right.php 730 2019-07-20 09:49:16Z webchills $
  */
 if (!defined('IS_ADMIN_FLAG')) {
   die('Illegal Access');
@@ -23,7 +23,7 @@ while (!$column_right_display->EOF and $box_cnt < 100) {
 ?>
 <?php
 //$column_box_spacer = 'column_box_spacer_right';
-$column_width = BOX_WIDTH_RIGHT;
+$column_width = (int)BOX_WIDTH_RIGHT;
 if ( file_exists(DIR_WS_MODULES . 'sideboxes/' . $template_dir . '/' . $column_right_display->fields['layout_box_name']) ) {
   $box_id = zen_get_box_id($column_right_display->fields['layout_box_name']);
   require(DIR_WS_MODULES . 'sideboxes/' . $template_dir . '/' . $column_right_display->fields['layout_box_name']);
